Pass first 3 tests
